requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
WTForms==3.1.2
SQLAlchemy==2.0.23
Flask-Reuploaded==1.4.0
Flask-Mail==0.9.1
email-validator== 2.1.1
gunicorn== 22.0.0
python-dotenv
flask_ckeditor
authlib

